Split tier budget guards out of the policy ladder step module (#256).
Unify skipped/completed step finish into one helper and move _TierStepRun plus clock helpers behind a focused budget module so the step runner stops absorbing unrelated concerns.
